Create Form
This is an offline based login/sign up form that allows users to signup/ login to an offline database. The login details of users are saved in a text editor in the project folder for the admin. It also has a log details for every user activity within the project. After a successful login, users can update details, change password, deactivate account, as well as log-out of the database. All activities of the user is thereby documented within the log book for an admin review.